+++ src/Intcode/IInput.cs
@@ -4,6 +4,6 @@ namespace Advent2019.Intcode
 {
     public interface IInput
     {
-        Task<int> GetNextAsync();
+        Task<long> GetNextAsync();
     }
 }
