+++ src/Intcode/IOutput.cs
@@ -4,6 +4,6 @@ namespace Advent2019.Intcode
 {
     public interface IOutput
     {
-        Task WriteAsync(int value);
+        Task WriteAsync(long value);
     }
 }
