+++ src/Intcode/Instructions/JumpIfFalse.cs
@@ -4,9 +4,9 @@ namespace Advent2019.Intcode.Instructions
 {
     public class JumpIfFalse : IInstruction
     {
-        public Task<int> RunAsync(Memory memory, int startingIndex, Command command, IInput input, IOutput output)
+        public Task<long> RunAsync(Memory memory, long startingIndex, Command command, IInput input, IOutput output)
         {
-            int a = memory.Read(startingIndex + 1, command.GetParameterMode(1));
+            long a = memory.Read(startingIndex + 1, command.GetParameterMode(1));
             if (a == 0)
             {
                 return Task.FromResult(memory.Read(startingIndex + 2, command.GetParameterMode(2)));
